refactor: cohesive state of BlockTimeDistribution chart
