+++ src/pages/StatisticsChart/index.tsx
@@ -25,7 +25,7 @@ import { UncleRateChart } from './mining/UncleRate'
 import { BalanceDistributionChart } from './activities/BalanceDistribution'
 import { TxFeeHistoryChart } from './activities/TxFeeHistory'
 import { BlockTimeDistributionChart } from './block/BlockTimeDistribution'
-import { getStatisticBlockTimeDistribution, getStatisticEpochTimeDistribution } from '../../service/app/charts/block'
+import { getStatisticEpochTimeDistribution } from '../../service/app/charts/block'
 import { EpochTimeDistributionChart } from './block/EpochTimeDistribution'
 import { NewDaoDepositChart } from './nervosDao/NewDaoDeposit'
 import { CirculationRatioChart } from './nervosDao/CirculationRatio'
@@ -245,7 +245,6 @@ export default () => {
     getStatisticTotalDaoDeposit(dispatch)
     getStatisticNewDaoDeposit(dispatch)
     getStatisticCirculationRatio(dispatch)
-    getStatisticBlockTimeDistribution(dispatch)
     getStatisticEpochTimeDistribution(dispatch)
     getStatisticTotalSupply(dispatch)
     getStatisticAnnualPercentageCompensation(dispatch)
