+++ src/service/app/charts/block.ts
@@ -1,53 +1,9 @@
-import { fetchStatisticBlockTimeDistribution, fetchStatisticEpochTimeDistribution } from '../../http/fetcher'
+import { fetchStatisticEpochTimeDistribution } from '../../http/fetcher'
 import { AppDispatch } from '../../../contexts/reducer'
 import { PageActions } from '../../../contexts/actions'
 import { fetchDateChartCache, storeDateChartCache } from '../../../utils/cache'
 import { ChartCachedKeys } from '../../../constants/cache'
-import { dispatchBlockTimeDistribution, dispatchEpochTimeDistribution } from './action'
-
-export const getStatisticBlockTimeDistribution = (dispatch: AppDispatch) => {
-  const data = fetchDateChartCache(ChartCachedKeys.BlockTimeDistribution)
-  if (data) {
-    dispatchBlockTimeDistribution(dispatch, data)
-    return
-  }
-  fetchStatisticBlockTimeDistribution()
-    .then((wrap: Response.Wrapper<State.StatisticBlockTimeDistributions> | null) => {
-      if (!wrap) return
-      const {
-        attributes: { blockTimeDistribution },
-      } = wrap
-      const sumBlocks = blockTimeDistribution
-        .flatMap(data => Number(data[1]))
-        .reduce((previous, current) => previous + current)
-      const statisticBlockTimeDistributions = [
-        {
-          time: '0',
-          ratio: '0',
-        },
-      ].concat(
-        blockTimeDistribution.map(data => {
-          const [time, blocks] = data
-          return {
-            time,
-            ratio: (Number(blocks) / sumBlocks).toFixed(5),
-          }
-        }),
-      )
-      dispatchBlockTimeDistribution(dispatch, statisticBlockTimeDistributions)
-      if (statisticBlockTimeDistributions && statisticBlockTimeDistributions.length > 0) {
-        storeDateChartCache(ChartCachedKeys.BlockTimeDistribution, statisticBlockTimeDistributions)
-      }
-    })
-    .catch(() => {
-      dispatch({
-        type: PageActions.UpdateStatisticBlockTimeDistributionFetchEnd,
-        payload: {
-          statisticBlockTimeDistributionsFetchEnd: true,
-        },
-      })
-    })
-}
+import { dispatchEpochTimeDistribution } from './action'
 
 export const getStatisticEpochTimeDistribution = (dispatch: AppDispatch) => {
   const data = fetchDateChartCache(ChartCachedKeys.EpochTimeDistribution)
